Move event-system code to `events` module
