+++ crates/wolf_engine_window/src/lib.rs
@@ -13,7 +13,10 @@
 //! # Handling Events, and Creating Window
 //!
 //! ```no_run
-//! # use wolf_engine_window::{WindowEvent, WindowSettings};
+//! # use wolf_engine_window::{
+//! #   WindowSettings,
+//! #   event::WindowEvent,
+//! # };
 //! #
 //! # let window_context = wolf_engine_window::init().build().unwrap();
 //! #
@@ -56,155 +59,26 @@ use std::{
 };
 
 use anyhow::anyhow;
-use error::{OsError, UnsupportedError, WindowError};
+use error::{OsError, WindowError};
+use event::EventLoopBuilder;
 use winit::{
     dpi::PhysicalSize,
-    error::EventLoopError,
-    event::{Event as WinitEvent, WindowEvent as WinitWindowEvent},
-    event_loop::{ActiveEventLoop, ControlFlow},
+    event_loop::ActiveEventLoop,
     window::{Window as WinitWindow, WindowAttributes, WindowId},
 };
 
-pub use uuid::Uuid;
+pub mod event;
 
-use wolf_engine_input::{Input, ToInput};
+pub use uuid::Uuid;
 
 /// Error-types used by the window system.
 pub mod error;
-
 /// Re-exports supported [`raw_window_handle`](crate::raw_window_handle::rwh_06) versions.
 pub mod raw_window_handle;
 
 /// Initialize the window system.
 pub fn init() -> EventLoopBuilder {
     EventLoopBuilder::new()
-}
-
-/// An event generated by the window system.
-#[derive(Clone, PartialEq)]
-#[non_exhaustive]
-pub enum WindowEvent {
-    Resumed,
-    RedrawRequested(Uuid),
-    Resized(Uuid, u32, u32),
-    Closed(Uuid),
-    Input(Option<Uuid>, Input),
-    Exited,
-}
-
-type WinitEventLoop = winit::event_loop::EventLoop<()>;
-
-/// Provides a way to configure the window system.
-///
-/// Create a new builder by calling [`init()`].
-pub struct EventLoopBuilder {
-    pub window_settings: WindowSettings,
-}
-
-impl EventLoopBuilder {
-    fn new() -> Self {
-        Self {
-            window_settings: WindowSettings::default(),
-        }
-    }
-
-    /// Initialize the window system.
-    pub fn build(self) -> Result<EventLoop, WindowError> {
-        match WinitEventLoop::with_user_event().build() {
-            Ok(event_loop) => Ok(self.build_with_event_loop(event_loop)),
-            Err(error) => match error {
-                EventLoopError::Os(error) => {
-                    Err(WindowError::OsError(OsError::from(anyhow!(error))))
-                }
-                EventLoopError::RecreationAttempt => Err(UnsupportedError::from(anyhow!(
-                    "Only 1 EventLoop can exist at a time"
-                ))
-                .into()),
-                error => panic!("Unhandled Error: {error}"),
-            },
-        }
-    }
-
-    #[allow(deprecated)]
-    fn build_with_event_loop(self, event_loop: WinitEventLoop) -> EventLoop {
-        EventLoop::new(event_loop)
-    }
-}
-
-/// The main-loop of the window system.
-pub struct EventLoop {
-    event_loop: WinitEventLoop,
-}
-
-impl EventLoop {
-    fn new(event_loop: WinitEventLoop) -> Self {
-        Self { event_loop }
-    }
-}
-
-impl EventLoop {
-    /// Run the main-loop, passing events to the provided callback.
-    #[allow(deprecated)]
-    pub fn run<F: FnMut(WindowEvent, &WindowContext)>(self, mut event_handler: F) {
-        let window_ids: Arc<RwLock<HashMap<WindowId, Uuid>>> =
-            Arc::new(RwLock::new(HashMap::new()));
-        let windows: Arc<RwLock<HashMap<Uuid, Weak<WinitWindow>>>> =
-            Arc::new(RwLock::new(HashMap::new()));
-        let _ = self.event_loop.run(|event, event_loop| {
-            let context = WindowContext::new(event_loop, window_ids.clone(), windows.clone());
-
-            match event {
-                WinitEvent::NewEvents(..) => {
-                    event_loop.set_control_flow(ControlFlow::Poll);
-
-                    windows
-                        .read()
-                        .expect("write-lock was acquired")
-                        .iter()
-                        .for_each(|(_, window)| {
-                            if let Some(window) = window.upgrade() {
-                                window.request_redraw();
-                            }
-                        });
-                }
-                WinitEvent::Resumed => (event_handler)(WindowEvent::Resumed, &context),
-                WinitEvent::LoopExiting => (event_handler)(WindowEvent::Exited, &context),
-                WinitEvent::DeviceEvent { event, .. } => {
-                    if let Some(input) = event.to_input() {
-                        (event_handler)(WindowEvent::Input(None, input), &context);
-                    }
-                }
-                WinitEvent::WindowEvent {
-                    window_id,
-                    event: window_event,
-                } => {
-                    let uuid = window_ids
-                        .read()
-                        .expect("read-lock was acquired")
-                        .get(&window_id)
-                        .expect("the uuid was inserted on window creation")
-                        .to_owned();
-                    if let Some(input) = window_event.to_input() {
-                        (event_handler)(WindowEvent::Input(Some(uuid), input), &context);
-                    }
-                    match window_event {
-                        WinitWindowEvent::RedrawRequested => {
-                            (event_handler)(WindowEvent::RedrawRequested(uuid), &context)
-                        }
-                        WinitWindowEvent::Resized(size) => (event_handler)(
-                            WindowEvent::Resized(uuid, size.width, size.height),
-                            &context,
-                        ),
-                        WinitWindowEvent::CloseRequested => {
-                            (event_handler)(WindowEvent::Closed(uuid), &context)
-                        }
-                        _ => (),
-                    }
-                }
-                _ => (),
-            }
-        });
-    }
 }
 
 /// A link to the window system.
@@ -400,6 +274,8 @@ mod window_init_tests {
     #[test]
     #[ntest::timeout(1000)]
     fn should_run_and_quit() {
+        use crate::event::{WindowEvent, WinitEventLoop};
+
         let event_loop = WinitEventLoop::with_user_event()
             .with_any_thread(true)
             .build()
