+++ examples/multiple_windows.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
 
 use pixels::{wgpu::Color, Pixels, SurfaceTexture};
-use wolf_engine::window::WindowEvent;
+use wolf_engine::window::event::WindowEvent;
 use wolf_engine_window::{Uuid, Window, WindowContext, WindowSettings};
 
 fn main() {
